requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
sqlalchemy
qrcode[pil]
Pillow
aiofiles
